Created ChatServiceImpl class in Server
In this class we get and return a StreamObserver response observer,via our method’s response observer while the client is still writing messages to their message stream.
Because we are using bidirectional streaming, both the client and server can read and write in any order — the streams operate completely independently.